requirements.txt
aws-cdk-lib
constructs
requests
kubernetes